effect_of_measurement_error.R
Effect on random (!) measurement error on prediction accuracy of the effect.
Low sample size results in an overestimation of the effect if the effect is small, and an underestimation of the effect if the effect is high without measurement error.
Measurement error increases the estimated effect if we only focus on p < 0.05 studies. This overestimation effect is kind of small while the measurement error is about half as big the effect itself. It gets more significant with larger measurement error.